fix:  통계 - 버그 수정 by 우영

+++ back/app/src/main/java/com/sloway/app/payment/stats/dto/response/PayMethodStatResDto.java
@@ -10,9 +10,9 @@ public class PayMethodStatResDto {
 
     private PayMethod payMethod;
     private Integer payCount;
-    private Integer totalAmt;
+    private Long totalAmt;   // 결제수단별 월 매출 — int 오버플로우 방지로 Long
 
-    public static PayMethodStatResDto of(PayMethod payMethod,Integer payCount ,Integer totalAmt) {
+    public static PayMethodStatResDto of(PayMethod payMethod, Integer payCount, Long totalAmt) {
         return PayMethodStatResDto.builder()
                 .payMethod(payMethod)
                 .payCount(payCount)
